Completing Random Food Suggestion Prototype
Added:
Scrolling using a ScrollView.
Recipe titles, pictures, and instructions that switch when the Random Recipe button at the bottom of the screen is pressed. (Implemented with TextSwitchers and ImageSwitchers)
Animations for when recipe info changes on button press.
Put our logo in the top left.
Imported several images to act as placeholder recipe pictures to the drawable folder.
Created several recipe instruction strings to act as placeholders in strings.xml

Changed:
The way recipe titles are displayed has been changed from a textView to a textSwitcher.