Update index.js - search functions
My currently deployed cloud functions:
Method <searchQuery> is used to search all of our user posts stored in our Firestore database. It takes an HTTPS put request that accepts a text string in the request body as the search query, such as "kimchi" or "Korean." If the request body text is empty, it will return all items from our Post collection in our Firestore. Returns a list of JSON objects as a string.
Method <searchTopHitOnly> is similar to searchQuery; it searches our Post collection in our Firestore DB, but it only returns the top hit form the search as a single JSON object in a string format.